update branch with main branch

+++ models/classModels.go
@@ -5,8 +5,8 @@ import (
 )
 
 type Class struct {
-	ClassID    uint   `gorm:"primaryKey" json:"id"`
-	TeacherID  uint   `json:"id_teacher" binding:"required"`
+	ClassID    int64  `gorm:"primaryKey" json:"id"`
+	TeacherID  int64  `json:"id_teacher" binding:"required"`
 	ClassName  string `json:"name" binding:"required"`
 	ClassGrade string `json:"grade" binding:"required"`
 	lib.BaseModel
